+++ relying-party/package.json
@@ -22,8 +22,6 @@
   },
   "dependencies": {
     "abstract-logging": "^2.0.1",
-    "ajv": "^8.11.0",
-    "ajv-formats": "^2.1.1",
     "jose": "^4.6.0",
     "lodash": "^4.17.21",
     "reflect-metadata": "^0.1.13",
