--- relying-party/src/Logger.ts
@@ -1,5 +1,0 @@
-export type LogLevels = "fatal" | "error" | "warn" | "info" | "debug" | "trace";
-
-export type AbstractLogging = {
-  [K in LogLevels]: (...args: any[]) => void;
-};
