--- relying-party/src/schemas/authentication-request.json
@@ -1,137 +1,0 @@
-{
-  "type": "object",
-  "properties": {
-    "client_id": {
-      "title": "Client Id",
-      "minLength": 1,
-      "maxLength": 2083,
-      "format": "uri",
-      "type": "string"
-    },
-    "response_type": {
-      "title": "Response Type",
-      "enum": ["code"],
-      "type": "string"
-    },
-    "scope": {
-      "type": "array",
-      "items": {
-        "$ref": "#/definitions/ScopeSpid"
-      }
-    },
-    "code_challenge": {
-      "title": "Code Challenge",
-      "type": "string"
-    },
-    "code_challenge_method": {
-      "title": "Code Challenge Method",
-      "enum": ["S256"],
-      "type": "string"
-    },
-    "nonce": {
-      "title": "Nonce",
-      "minLength": 32,
-      "type": "string"
-    },
-    "redirect_uri": {
-      "title": "Redirect Uri",
-      "minLength": 1,
-      "maxLength": 2083,
-      "format": "uri",
-      "type": "string"
-    },
-    "claims": {
-      "title": "Claims",
-      "type": "object"
-    },
-    "state": {
-      "title": "State",
-      "minLength": 32,
-      "type": "string"
-    },
-    "ui_locales": {
-      "title": "Ui Locales",
-      "type": "array",
-      "items": {
-        "type": "string"
-      }
-    },
-    "sub": {
-      "title": "Sub",
-      "minLength": 1,
-      "maxLength": 2083,
-      "format": "uri",
-      "type": "string"
-    },
-    "iss": {
-      "title": "Iss",
-      "minLength": 1,
-      "maxLength": 2083,
-      "format": "uri",
-      "type": "string"
-    },
-    "iat": {
-      "title": "Iat",
-      "type": "integer"
-    },
-    "exp": {
-      "title": "Exp",
-      "type": "integer"
-    },
-    "jti": {
-      "title": "Jti",
-      "type": "string"
-    },
-    "aud": {
-      "title": "Aud",
-      "type": "array",
-      "items": {
-        "type": "string",
-        "minLength": 1,
-        "maxLength": 2083,
-        "format": "uri"
-      }
-    },
-    "acr_values": {
-      "type": "array",
-      "items": {
-        "$ref": "#/definitions/AcrValues"
-      }
-    },
-    "prompt": {
-      "title": "Prompt",
-      "enum": ["consent", "consent login", "verify"],
-      "type": "string"
-    }
-  },
-  "required": [
-    "client_id",
-    "response_type",
-    "scope",
-    "code_challenge",
-    "code_challenge_method",
-    "nonce",
-    "redirect_uri",
-    "state",
-    "sub",
-    "iss",
-    "iat",
-    "aud",
-    "acr_values",
-    "prompt"
-  ],
-  "definitions": {
-    "ScopeSpid": {
-      "title": "ScopeSpid",
-      "description": "An enumeration.",
-      "enum": ["openid", "offline_access"],
-      "type": "string"
-    },
-    "AcrValues": {
-      "title": "AcrValues",
-      "description": "An enumeration.",
-      "enum": ["https://www.spid.gov.it/SpidL1", "https://www.spid.gov.it/SpidL2", "https://www.spid.gov.it/SpidL3"],
-      "type": "string"
-    }
-  }
-}
